some refactoring of signing
