+++ rust/agents/validator/src/submit.rs
@@ -7,7 +7,7 @@ use tokio::{task::JoinHandle, time::sleep};
 use tracing::{debug, info, info_span, instrument::Instrumented, Instrument};
 
 use hyperlane_base::{CachingMailbox, CheckpointSyncer, CheckpointSyncers, CoreMetrics};
-use hyperlane_core::{Announcement, HyperlaneDomain, HyperlaneSigner, Mailbox, Signable};
+use hyperlane_core::{Announcement, HyperlaneDomain, HyperlaneSigner, HyperlaneSignerExt, Mailbox};
 
 pub(crate) struct ValidatorSubmitter {
     interval: u64,
@@ -55,7 +55,7 @@ impl ValidatorSubmitter {
             mailbox_domain: self.mailbox.mailbox().domain().id(),
             storage_metadata: self.checkpoint_syncer.announcement_metadata(),
         };
-        let signed_announcement = announcement.sign_with(&self.signer).await?;
+        let signed_announcement = self.signer.sign(announcement).await?;
         self.checkpoint_syncer
             .write_announcement(&signed_announcement)
             .await?;
@@ -132,7 +132,7 @@ impl ValidatorSubmitter {
                 .map(|i| i < latest_checkpoint.index)
                 .unwrap_or(true)
             {
-                let signed_checkpoint = latest_checkpoint.sign_with(&self.signer).await?;
+                let signed_checkpoint = self.signer.sign(latest_checkpoint).await?;
 
                 info!(signed_checkpoint = ?signed_checkpoint, signer=?self.signer, "Signed new latest checkpoint");
                 current_index = Some(latest_checkpoint.index);
